Typo refactor package name
