+++ src/main/java/id/my/btw/repository/ExpenseRepository.java
@@ -1,6 +1,6 @@
-package id.btw.repository;
+package id.my.btw.repository;
 
-import id.btw.entity.Expense;
+import id.my.btw.entity.Expense;
 
 public interface ExpenseRepository {
     void insert(Expense expense);
